+++ bluefin-contrib/CHANGELOG.md
@@ -1,1 +1,3 @@
 # 0.16.0.0
+
+* Add `runManyStates`, thanks to Kari Pahula